requirements/docs.txt
# Requirements for docs.
autodoc-pydantic
furo
Sphinx
sphinx-autobuild
sphinx-click
